changed the file structure as requested
