+++ app/about/page.tsx
@@ -1,8 +1,10 @@
-import React from "react";
-import Navbar from "../../components/Navbar";
+import React from 'react'
+import About from '@/components/about/page'
 
-const About = () => {
-  return <div>This is the about page here</div>;
-};
+const aboutpage = () => {
+  return (
+    <About />
+  )
+}
 
-export default About;
+export default aboutpage;
