+++ app/globals.css
@@ -99,7 +99,7 @@ body {
   height: 100vh;
 }
 .background-video {
-  position: fixed;
+  position: absolute;
   top: 0;
   left: 0;
   min-width: 100%;
