+++ components/Footer.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <div className='flex gap-2 justify-between items-center'>
+        <div>
+            Lorem ipsum dolor sit amet consectetur, adipisicing elit. Iure, quis quibusdam iste quas nobis eaque distinctio dicta similique, quidem suscipit dolore veniam nihil doloribus pariatur reprehenderit odit odio quaerat vel et aliquid fugit. Commodi voluptas sed nihil fuga sit a accusantium cum ex nisi praesentium facere quo iure optio, est quod modi reprehenderit omnis magnam. Dolorum doloribus blanditiis tempore quia quidem modi provident odio dolore labore omnis sint cupiditate officia, a fugiat dolores. Neque quia id nulla a, soluta ullam tenetur maxime distinctio commodi cupiditate. Corporis vitae minima harum esse nesciunt, doloremque facilis voluptate, quidem quibusdam aperiam aut, odio incidunt!
+        </div>
+    </div>
+  )
+}
+
+export default Footer
