+++ components/about/page.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Navbar from "../../components/Navbar";
+
+const About = () => {
+  return <div>This is the about page here</div>;
+};
+
+export default About;
