+++ components/contact/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Contact = () => {
+  return (
+    <div>This is the contact us page</div>
+  )
+}
+
+export default Contact
