+++ components/events/page.tsx
@@ -1,0 +1,10 @@
+'use client'
+import React from 'react'
+
+const Events = () => {
+  return (
+    <div>page</div>
+  )
+}
+
+export default Events
