+++ components/signin/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SignIn = () => {
+  return (
+    <div>This is where you sign up or sign in</div>
+  )
+}
+
+export default SignIn
